requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
openai==1.42.0
fal_client==0.4.1
python-dotenv